feat(i18n): introduce translation hooks and formatMessage utility - Added useTranslator hook for simplified translation function access. - Implemented formatMessage utility for enhanced message formatting with context. - Updated index files to export new hooks and utilities.
