+++ src/libs/i18n/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useTranslator'
