+++ src/libs/i18n/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './translator'
+export * from './formatMessage'
 export * from './models'
